base case of type inference recursion in place
git-svn-id: https://stan.googlecode.com/svn/trunk@200 9a304bd9-dce1-f7c0-8d5c-bb1642157d4e

+++ src/stan/io/dump.hpp
@@ -18,7 +18,7 @@ namespace stan {
   namespace io {
 
      namespace {
-       double product(std::vector<unsigned int> dims) {
+       unsigned int product(std::vector<unsigned int> dims) {
 	 unsigned int y = 1U;
 	 for (unsigned int i = 0; i < dims.size(); ++i)
 	   y *= dims[i];
